+++ webapp/src/pages/Game.js
@@ -4,13 +4,15 @@ import axios from "axios";
 import { useLocation } from 'react-router-dom';
 
 function Game() {
-  const QUESTION_TIME = 50;
+  const QUESTION_TIME = 10;
+  const TOTAL_ROUNDS = 10;
 
   const location = useLocation();
 
   const [score, setScore] = useState(0);
   const [timeLeft, setTimeLeft] = useState(QUESTION_TIME);
   const [gameMode, setGameMode] = useState('');
+  const [round, setRound] = useState(1);
 
   const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
 
@@ -32,7 +34,7 @@ function Game() {
 
   const fetchQuestion = async () => {
     try {
-      if (!gameMode) return;
+      if (!gameMode || round > TOTAL_ROUNDS) return;
       const response = await axios.get(`${apiEndpoint}/questions/${gameMode}`);
       setQuestionData(response.data); 
     } catch (error) {
@@ -45,11 +47,13 @@ function Game() {
       fetchQuestion(); // Llama a la función para obtener la pregunta solo si gameMode está definido
     }
   }, [gameMode]); 
+
   // Función para manejar las respuestas
   const handleAnswer = (isCorrect) => {
     if (isCorrect) {
       setScore(score + 1);
     }
+    setRound(prevRound => prevRound + 1);
     fetchQuestion();
     setTimeLeft(QUESTION_TIME);
   };
@@ -87,6 +91,21 @@ function Game() {
         }}
       >
         <Typography variant="h6">Puntuación: {score}</Typography>
+      </Box>
+
+      {/* Round counter */}
+      <Box
+        sx={{
+          position: "absolute",
+          top: "5vh", 
+          left: "50%", 
+          transform: "translateX(-50%)", // Centra el contador de rondas
+          padding: "0.5vw",
+          border: "2px solid green",
+          borderRadius: "0.5vw",
+        }}
+      >
+      <Typography variant="h6">Ronda {round} de {TOTAL_ROUNDS}</Typography>
       </Box>
 
       {/* Pregunta sobre la imagen */}
