+++ webapp/src/pages/PrivateRoute.js
@@ -4,7 +4,6 @@ import { Navigate } from 'react-router-dom';
 const PrivateRoute = ({ element: Element }) => {
     const storedSessionId = localStorage.getItem('sessionId');
  
-
     return ( storedSessionId ? <Element /> : <Navigate to="/login" /> );
 
 };
